feat(util): 收敛通用工具行为与兼容性覆盖
- 重构 `util/circuit`、`util/retry`、`util/config`、`util/encoding`、`util/env`、`util/file` 的核心实现，收敛并发关闭、错误链、配置解析和安全行为
- 增补 `util/hash`、`util/idgen`、`util/json`、`util/lease`、`util/logger`、`util/pagination`、`util/poolx`、`util/rand`、`util/rate`、`util/reflectx`、`util/slice`、`util/validator` 的回归与 fuzz/covenant 测试
- 清理 `util/hash/coverage.out` 产物，避免覆盖率文件污染提交树
- 兼顾稳定性与兼容性，不添加旧废弃 API，保持新 API 契约为主

影响范围：本次提交覆盖 toolkit 通用库大规模行为收敛，理论上可能影响依赖这些 util 的上游模块运行时细节（重试、并发和日志语义），未新增对外 API 名称。

+++ util/pagination/pagination_test.go
@@ -7,12 +7,12 @@ import (
 func TestNew(t *testing.T) {
 	tests := []struct {
 		name       string
-		page       int
+		page       int64
 		pageSize   int
 		total      int64
-		wantPage   int
-		wantOffset int
-		wantPages  int
+		wantPage   int64
+		wantOffset int64
+		wantPages  int64
 	}{
 		{"normal", 1, 10, 100, 1, 0, 10},
 		{"page 2", 2, 10, 100, 2, 10, 10},
